Fixed new accounts not being handled properly in Game
The game method was listening for an exception that was being handled in the PlayerDataHandler, so it skipped the catch block in the Game class, causing the error to not be handled properly. This caused the player object to be passed as null when it was not loaded instead of passing a new player.